Delete Hardhat (#372)
