+++ .github/workflows/tests.yml
@@ -7,30 +7,6 @@ on:
   pull_request:
 
 jobs:
-  unit-tests:
-    name: Hardhat Unit Tests
-    runs-on: ubuntu-latest
-
-    steps:
-      - uses: actions/checkout@v3
-        with:
-          submodules: recursive
-
-      - uses: actions/setup-node@v3
-        with:
-          node-version: 16
-          cache: 'yarn'
-
-      - name: Install dependencies
-        run: yarn install --frozen-lockfile
-
-      # This is required separately from yarn test because it generates the typechain definitions
-      - name: Compile
-        run: yarn compile
-
-      - name: Run unit tests
-        run: yarn test
-
   forge-tests:
     name: Forge Tests
     runs-on: ubuntu-latest
