--- hardhat.config.ts
@@ -1,51 +1,0 @@
-import '@nomiclabs/hardhat-ethers'
-import '@nomiclabs/hardhat-etherscan'
-import '@nomiclabs/hardhat-waffle'
-import 'hardhat-typechain'
-
-const importToml = require('import-toml')
-const foundryConfig = importToml.sync('foundry.toml')
-
-export default {
-  networks: {
-    hardhat: {
-      allowUnlimitedContractSize: true,
-    },
-    mainnet: {
-      url: `https://mainnet.infura.io/v3/${process.env.INFURA_API_KEY}`,
-    },
-    ropsten: {
-      url: `https://ropsten.infura.io/v3/${process.env.INFURA_API_KEY}`,
-    },
-    rinkeby: {
-      url: `https://rinkeby.infura.io/v3/${process.env.INFURA_API_KEY}`,
-    },
-    goerli: {
-      url: `https://goerli.infura.io/v3/${process.env.INFURA_API_KEY}`,
-    },
-    kovan: {
-      url: `https://kovan.infura.io/v3/${process.env.INFURA_API_KEY}`,
-    },
-  },
-  etherscan: {
-    // Your API key for Etherscan
-    // Obtain one at https://etherscan.io/
-    apiKey: process.env.ETHERSCAN_API_KEY,
-  },
-  solidity: {
-    version: foundryConfig.profile.default.solc_version,
-    settings: {
-      viaIR: foundryConfig.profile.default.via_ir,
-      optimizer: {
-        enabled: true,
-        runs: foundryConfig.profile.default.optimizer_runs,
-      },
-      metadata: {
-        // do not include the metadata hash, since this is machine dependent
-        // and we want all generated code to be deterministic
-        // https://docs.soliditylang.org/en/v0.8.20/metadata.html
-        bytecodeHash: 'none',
-      },
-    },
-  },
-}
