+++ justfile
@@ -1,30 +1,18 @@
-test: test-forge test-hardhat
+test: test-forge
 prep: fix snapshots
-snapshots: snapshots-forge snapshots-hardhat
+snapshots: snapshots-forge
 
 test-forge: install-forge build-forge
     forge test
 
-test-hardhat: install-hardhat
-    yarn test
-
 build-forge: install-forge
     forge build
-
-build-hardhat: install-hardhat
-    yarn build
 
 snapshots-forge: install-forge test-forge
     forge snapshot
 
-snapshots-hardhat: install-hardhat
-    yarn snapshots
-
 install-forge:
     forge install
-
-install-hardhat:
-    yarn install
 
 fix:
     forge fmt
