--- test/shared/constants.ts
@@ -1,4 +1,0 @@
-export const MIN_TICK = -887272
-export const MAX_TICK = 887272
-export const MAX_TICK_SPACING = 32767
-export const ADDRESS_ZERO = '0x0000000000000000000000000000000000000000'
