--- test/shared/expect.ts
@@ -1,8 +1,0 @@
-import { expect, use } from 'chai'
-import { solidity } from 'ethereum-waffle'
-import { jestSnapshotPlugin } from 'mocha-chai-jest-snapshot'
-
-use(solidity)
-use(jestSnapshotPlugin())
-
-export { expect }
