--- test/shared/fixtures.ts
@@ -1,22 +1,0 @@
-import { BigNumber } from 'ethers'
-import { ethers } from 'hardhat'
-import { TestERC20 } from '../../typechain/TestERC20'
-
-interface TokensFixture {
-  currency0: TestERC20
-  currency1: TestERC20
-  token2: TestERC20
-}
-
-export async function tokensFixture(): Promise<TokensFixture> {
-  const tokenFactory = await ethers.getContractFactory('TestERC20')
-  const tokenA = (await tokenFactory.deploy(BigNumber.from(2).pow(255))) as TestERC20
-  const tokenB = (await tokenFactory.deploy(BigNumber.from(2).pow(255))) as TestERC20
-  const tokenC = (await tokenFactory.deploy(BigNumber.from(2).pow(255))) as TestERC20
-
-  const [currency0, currency1, token2] = [tokenA, tokenB, tokenC].sort((tokenA, tokenB) =>
-    tokenA.address.toLowerCase() < tokenB.address.toLowerCase() ? -1 : 1
-  )
-
-  return { currency0, currency1, token2 }
-}
