+++ tsconfig.json
@@ -6,8 +6,7 @@
     "esModuleInterop": true,
     "outDir": "dist",
     "typeRoots": ["./typechain", "./node_modules/@types"],
-    "types": ["@nomiclabs/hardhat-ethers", "@nomiclabs/hardhat-waffle"]
+    "types": ["@nomiclabs/hardhat-ethers"]
   },
   "include": ["./test"],
-  "files": ["./hardhat.config.ts"]
 }
